+++ hmda-data-publisher/src/main/scala/hmda/publisher/query/component/PublisherComponent2018.scala
@@ -30,7 +30,7 @@ trait PublisherComponent2018 {
     def respondentCity = column[String]("respondent_city")
     def parentIdRssd = column[Int]("parent_id_rssd")
     def parentName = column[String]("parent_name")
-    def assets = column[Int]("assets")
+    def assets = column[Long]("assets")
     def otherLenderCode = column[Int]("other_lender_code")
     def topHolderIdRssd = column[Int]("topholder_id_rssd")
     def topHolderName = column[String]("topholder_name")
